services/core_service/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
asyncpg
pydantic
pydantic-settings
python-jose[cryptography]
python-multipart
aiofiles
chromadb
google-genai
httpx